+++ libs/@guardian/atoms-rendering/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @guardian/atoms-rendering
 
+## 27.0.0
+
+### Major Changes
+
+- Now requires TypeScript ~4.9.5 (if you use it with TypeScript).
+
+### Patch Changes
+
+- Updated dependencies
+  - @guardian/libs@14.0.0
+  - @guardian/source-foundations@10.0.0
+  - @guardian/source-react-components@12.0.0
+  - @guardian/eslint-plugin-source-foundations@11.0.0
+  - @guardian/eslint-plugin-source-react-components@14.0.0
+
 ## 26.0.0
 
 ### Major Changes