+++ libs/@guardian/source-react-components/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @guardian/source-react-components
 
+## 12.0.0
+
+### Major Changes
+
+- Now requires TypeScript ~4.9.5 (if you use it with TypeScript).
+
+### Patch Changes
+
+- Updated dependencies
+  - @guardian/source-foundations@10.0.0
+
+## 11.4.0
+
+### Minor Changes
+
+- 41095ce: Adds optional prop to CheckboxGroup element to append "optional" text to label
+
 ## 11.3.0
 
 ### Minor Changes